refactor: convert dashboard page to server component
